pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claudium"
version = "0.1.0"
description = "The agent harness framework for Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
    "python-frontmatter>=1.1",
    "tomli>=2.0; python_version < '3.11'",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "sse-starlette>=2.0",
]
sandboxes = [
    "e2b>=0.17",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
claudium = "claudium.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
